Added calendar files processing

+++ transit-routing-backend/src/main/java/com/dnikitin/transit/infrastructure/repository/ServiceCalendarJpaRepository.java
@@ -3,5 +3,5 @@ package com.dnikitin.transit.infrastructure.repository;
 import com.dnikitin.transit.infrastructure.persistence.entity.ServiceCalendarEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceCalendarJpaRepository extends JpaRepository<ServiceCalendarEntity, Integer> {
+public interface ServiceCalendarJpaRepository extends JpaRepository<ServiceCalendarEntity, Long> {
 }
